// src/data/node-types/magazine-issues/types/ApiMagazineIssueDocumentsLapTimeRelationship.ts
import type {ApiRelationshipType} from "../../../types/ApiRelationshipType"
import type {ApiNodeType} from "../../../types/ApiNodeType"
import type {ApiLapTimeNode} from "../../lap-times/types/ApiLapTimeNode"

export type ApiMagazineIssueDocumentsLapTimeRelationship = {
    links: {
        self: string
    }
    data: {
        type: ApiNodeType.LAP_TIME
        id: number
        attributes: Record<string, string | number | boolean | null>
        data: {
            relationship_id: number
            relationship_name: ApiRelationshipType.MAGAZINE_ISSUE_DOCUMENTS_LAP_TIME
            partner_node: {
                node_type: ApiNodeType.LAP_TIME
                data: ApiLapTimeNode['attributes']
            }
            created_at: string
            updated_at: string
        }
    }[]
}